fix: Fixes connection pool minimum size to 0, also refactored uri related operations
