+++ app/server/appsmith-plugins/mongoPlugin/src/test/java/com/external/plugins/MongoPluginTest.java
@@ -22,13 +22,13 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.mongodb.DBRef;
 import com.mongodb.MongoCommandException;
 import com.mongodb.MongoSecurityException;
 import com.mongodb.reactivestreams.client.MongoClient;
 import com.mongodb.reactivestreams.client.MongoClients;
 import com.mongodb.reactivestreams.client.MongoCollection;
 import com.mongodb.reactivestreams.client.MongoDatabase;
-import com.mongodb.DBRef;
 import org.bson.Document;
 import org.bson.types.BSONTimestamp;
 import org.bson.types.Decimal128;
@@ -75,6 +75,7 @@ import static com.external.plugins.constants.FieldName.SMART_SUBSTITUTION;
 import static com.external.plugins.constants.FieldName.UPDATE_LIMIT;
 import static com.external.plugins.constants.FieldName.UPDATE_OPERATION;
 import static com.external.plugins.constants.FieldName.UPDATE_QUERY;
+import static com.external.plugins.utils.DatasourceUtils.buildClientURI;
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -2432,59 +2433,6 @@ public class MongoPluginTest {
                     assertEquals(value.asInt(), 3);
                 })
                 .verifyComplete();
-    }
-
-    @Test
-    public void testBuildClientURI_withoutUserInfoAndAuthSource() {
-
-        final String testUri = "mongodb://host:port/db?param";
-        final String resultUri = "mongodb://host:port/db?param&authSource=admin";
-
-        DatasourceConfiguration datasourceConfiguration = new DatasourceConfiguration();
-        final DBAuth dbAuth = new DBAuth();
-        datasourceConfiguration.setAuthentication(dbAuth);
-        datasourceConfiguration.setProperties(List.of(
-                new Property("0", "Yes"),
-                new Property("1", testUri)
-        ));
-        final String clientURI = pluginExecutor.buildClientURI(datasourceConfiguration);
-        assertEquals(resultUri, clientURI);
-    }
-
-    @Test
-    public void testBuildClientURI_withUserInfoAndAuthSource() {
-
-        final String testUri = "mongodb://user:pass@host:port/db?param&authSource=notAdmin";
-        final String resultUri = "mongodb://user:newPass@host:port/db?param&authSource=notAdmin";
-
-        DatasourceConfiguration datasourceConfiguration = new DatasourceConfiguration();
-        final DBAuth dbAuth = new DBAuth();
-        dbAuth.setPassword("newPass");
-        datasourceConfiguration.setAuthentication(dbAuth);
-        datasourceConfiguration.setProperties(List.of(
-                new Property("0", "Yes"),
-                new Property("1", testUri)
-        ));
-        final String clientURI = pluginExecutor.buildClientURI(datasourceConfiguration);
-        assertEquals(resultUri, clientURI);
-    }
-
-    @Test
-    public void testBuildClientURI_withoutDbInfoAndPortsAndParams() {
-
-        final String testUri = "mongodb://user:pass@host";
-        final String resultUri = "mongodb://user:newPass@host/?authSource=admin";
-
-        DatasourceConfiguration datasourceConfiguration = new DatasourceConfiguration();
-        final DBAuth dbAuth = new DBAuth();
-        dbAuth.setPassword("newPass");
-        datasourceConfiguration.setAuthentication(dbAuth);
-        datasourceConfiguration.setProperties(List.of(
-                new Property("0", "Yes"),
-                new Property("1", testUri)
-        ));
-        final String clientURI = pluginExecutor.buildClientURI(datasourceConfiguration);
-        assertEquals(resultUri, clientURI);
     }
 
     @Test
